Move git functions to git_commands, make commit after updating icons
